Moved admin code from init to separate file
All code for adminpanel has been moved from __init__.py to a new,
separate file, admin.py.
